update to ember-cli add-on structure
