+++ .bowerrc
@@ -1,0 +1,4 @@
+{
+  "directory": "bower_components",
+  "analytics": false
+}
