+++ .travis.yml
@@ -1,7 +1,34 @@
-script: bundle exec rake test
-after_success: bundle exec rake publish_build
+---
+language: node_js
+node_js:
+  - "0.12"
+
+sudo: false
+
+cache:
+  directories:
+    - node_modules
+
 env:
-  global:
-  - S3_BUCKET_NAME=builds.emberjs.com
-  - secure: RWe9Wsc1/uCFiAPxTp+tnsx37BTlCNGvObsZzF3FS8hb0uSIz79Eou5bPT5cn4uiDR71GW/XIoXa9RjkqHFNdNiZqrG5YOC/VCfDxKZ81OoYlKJwrhm9ats8REGzFXMdW6VHxh/a1IlPk2iQh133uaLSEbem7Py1AJVhL16FuSI=
-  - secure: dfZXMaz+1S3Xdn6ImyuBMlb2C1NULkqljTYb0wmvL7gmO2vluLuLq7zikbci8Nw4q8ckg25jHdIBwesrF1WK9ls/+k0elA8sYZNmy5PMghWevmn86PtUog1cOtulstVO2gnjFQnQNG4/pm4UX/JzhXSWpjCMEU+0ibQ3J4eSyJM=
+  - EMBER_TRY_SCENARIO=default
+  - EMBER_TRY_SCENARIO=ember-release
+  - EMBER_TRY_SCENARIO=ember-beta
+  - EMBER_TRY_SCENARIO=ember-canary
+
+matrix:
+  fast_finish: true
+  allow_failures:
+    - env: EMBER_TRY_SCENARIO=ember-canary
+
+before_install:
+  - export PATH=/usr/local/phantomjs-2.0.0/bin:$PATH
+  - "npm config set spin false"
+  - "npm install -g npm@^2"
+
+install:
+  - npm install -g bower
+  - npm install
+  - bower install
+
+script:
+  - ember try $EMBER_TRY_SCENARIO test
