--- config.ru
@@ -1,5 +1,0 @@
-require 'bundler/setup'
-require 'ember/source'
-require 'ember-dev'
-
-run EmberDev::Server.new
