--- docs/README.md
@@ -1,20 +1,0 @@
-Ember States Docs
-=================
-
-Basic YuiDoc documentation for Ember-States.
-
-### Build 
-
-Install dependencies:
-
-```
-npm install -g git://github.com/wagenet/yuidoc.git
-```
-
-Generate Docs:
-
-```
-yuidoc
-```
-
-The generate documentation should be located in the `build` directory.
