+++ dummy/app/templates/application.hbs
@@ -1,0 +1,3 @@
+<h2 id="title">Welcome to Ember.js</h2>
+
+{{outlet}}
