+++ dummy/public/robots.txt
@@ -1,0 +1,3 @@
+# http://www.robotstxt.org
+User-agent: *
+Disallow:
