--- ember-dev.yml
@@ -1,16 +1,0 @@
----
-name: Ember States
-assetfile: 'Assetfile'
-testing_suites:
-  default:
-    - "package=all"
-  all:
-    - "package=all"
-    - "package=all&jquery=1.7.2&nojshint=true"
-    - "package=all&jquery=1.8.3&nojshint=true"
-    - "package=all&jquery=1.9.1&nojshint=true"
-    - "package=all&extendprototypes=true&nojshint=true"
-    - "package=all&extendprototypes=true&jquery=1.9.1&nojshint=true"
-    - "package=all&dist=build&nojshint=true"
-testing_packages:
-  - ember-states
