+++ index.js
@@ -1,0 +1,6 @@
+/* jshint node: true */
+'use strict';
+
+module.exports = {
+  name: 'ember-states'
+};
