+++ tests/.jshintrc
@@ -2,13 +2,33 @@
   "predef": [
     "document",
     "window",
-    "-Promise"
+    "location",
+    "setTimeout",
+    "$",
+    "-Promise",
+    "define",
+    "console",
+    "visit",
+    "exists",
+    "fillIn",
+    "click",
+    "keyEvent",
+    "triggerEvent",
+    "find",
+    "findWithAssert",
+    "wait",
+    "DS",
+    "andThen",
+    "currentURL",
+    "currentPath",
+    "currentRouteName"
   ],
-  "browser": true,
+  "node": false,
+  "browser": false,
   "boss": true,
   "curly": true,
   "debug": false,
-  "devel": true,
+  "devel": false,
   "eqeqeq": true,
   "evil": true,
   "forin": false,
