--- tests/ember_configuration.js
@@ -1,25 +1,0 @@
-/*globals EmberDev ENV QUnit */
-
-(function() {
-  window.Ember = window.Ember || {};
-
-  Ember.config = {};
-  Ember.testing = true;
-
-  window.ENV = { TESTING: true };
-
-  var extendPrototypes = QUnit.urlParams.extendprototypes;
-  ENV['EXTEND_PROTOTYPES'] = !!extendPrototypes;
-
-  if (EmberDev.jsHint) {
-    // jsHint makes its own Object.create stub, we don't want to use this
-    ENV['STUB_OBJECT_CREATE'] = !Object.create;
-  }
-
-  EmberDev.distros = {
-    //spade:   'ember-data-spade.js',
-    spade:   'ember-spade.js',
-    build:   'ember-data.js'
-  };
-
-})();
